feat(blocks): Add lego arch bricks

+++ core/src/commonMain/kotlin/org/micoli/micraft/game/world/BlockDefinition.kt
@@ -18,7 +18,7 @@ data class BlockDefinition(
     val minimapVisible: Boolean = true,
     val rotatable: Boolean = false,
     val hasStuds: Boolean = false,
-    val brickSize: List<Int> = listOf(1, 1, 1),
+    val brickSize: List<Float> = listOf(1f, 1f, 1f),
     val heightFraction: Float = 1.0f,
     /** When true, the block can be placed in any palette color (see [PlainColorRegistry]). */
     val plainColorable: Boolean = false,
